Reverted to edit edt-4f398011-a30b-49ee-98b9-bd3447423206: "Fix: Supabase authentication error
The Supabase authentication is failing with a "Failed to fetch" error.  This commit aims to resolve the underlying issue causing the authentication failure."
